+++ savvy-bindgen/src/lib.rs
@@ -1,13 +1,12 @@
+mod gen;
 mod parse_file;
 mod savvy_fn;
 mod savvy_impl;
 mod utils;
 
-pub use savvy_fn::{
-    generate_c_header_file, generate_c_impl_file, generate_r_impl_file, ParsedResult, SavvyFn,
-    SavvyFnArg, SavvyFnType,
-};
-
+pub use gen::c::{generate_c_header_file, generate_c_impl_file};
+pub use gen::r::generate_r_impl_file;
+pub use savvy_fn::{ParsedResult, SavvyFn, SavvyFnArg, SavvyFnType};
 pub use savvy_impl::SavvyImpl;
 
 pub use utils::extract_docs;
