+++ savvy-bindgen/src/savvy_fn.rs
@@ -87,6 +87,28 @@ pub struct SavvyFnArg {
     ty: SavvySupportedTypes,
 }
 
+impl SavvyFnArg {
+    pub fn pat(&self) -> syn::Ident {
+        self.pat.clone()
+    }
+
+    pub fn pat_string(&self) -> String {
+        self.pat.to_string()
+    }
+
+    pub fn to_c_type_string(&self) -> String {
+        self.ty.to_c_type()
+    }
+
+    pub fn to_rust_type_outer(&self) -> syn::Type {
+        self.ty.to_rust_type_outer()
+    }
+
+    pub fn to_rust_type_inner(&self) -> syn::Type {
+        self.ty.to_rust_type_inner()
+    }
+}
+
 pub enum SavvyFnType {
     BareFunction,
     Constructor(syn::Type),
@@ -115,7 +137,7 @@ pub struct SavvyFn {
 
 #[allow(dead_code)]
 impl SavvyFn {
-    fn get_self_ty_ident(&self) -> Option<syn::Ident> {
+    pub(crate) fn get_self_ty_ident(&self) -> Option<syn::Ident> {
         let self_ty = match &self.fn_type {
             SavvyFnType::BareFunction => return None,
             SavvyFnType::Constructor(ty) => ty,
@@ -147,21 +169,6 @@ impl SavvyFn {
         match self.get_self_ty_ident() {
             Some(ty) => format_ident!("savvy_{}_{}", ty, self.fn_name),
             None => format_ident!("savvy_{}", self.fn_name),
-        }
-    }
-
-    pub fn fn_name_r(&self) -> syn::Ident {
-        match self.get_self_ty_ident() {
-            Some(ty) => {
-                // Special convention. If the method name is "new", use type
-                // itself.
-                if self.fn_name.to_string().as_str() == "new" {
-                    ty
-                } else {
-                    format_ident!("{}_{}", ty, self.fn_name)
-                }
-            }
-            None => self.fn_name.clone(),
         }
     }
 
@@ -244,500 +251,4 @@ impl SavvyFn {
             stmts_additional,
         }
     }
-
-    pub fn generate_inner_fn(&self) -> syn::ItemFn {
-        let fn_name_orig = &self.fn_name;
-        let fn_name_inner = self.fn_name_inner();
-
-        let args_pat: Vec<syn::Ident> = self.args.iter().map(|arg| arg.pat.clone()).collect();
-        let args_ty: Vec<syn::Type> = self
-            .args
-            .iter()
-            .map(|arg| arg.ty.to_rust_type_inner())
-            .collect();
-
-        let stmts_additional = &self.stmts_additional;
-        let stmts_orig = &self.stmts_orig;
-        let attrs = &self.attrs;
-
-        let out: syn::ItemFn = match (&self.fn_type, self.has_result) {
-            // A bare function with result
-            (SavvyFnType::BareFunction, true) => parse_quote!(
-                #(#attrs)*
-                unsafe fn #fn_name_inner( #(#args_pat: #args_ty),* ) -> savvy::Result<savvy::SEXP> {
-                    #(#stmts_additional)*
-                    #(#stmts_orig)*
-                }
-            ),
-            // A bare function without result; return a dummy value
-            (SavvyFnType::BareFunction, false) => parse_quote!(
-                #(#attrs)*
-                unsafe fn #fn_name_inner( #(#args_pat: #args_ty),* ) -> savvy::Result<savvy::SEXP> {
-                    #(#stmts_additional)*
-                    #(#stmts_orig)*
-
-                    // Dummy return value
-                    Ok(savvy::NullSxp.into())
-                }
-            ),
-            // A method with result
-            (SavvyFnType::Method(ty), true) => parse_quote!(
-                #(#attrs)*
-                #[allow(non_snake_case)]
-                unsafe fn #fn_name_inner(self__: savvy::SEXP, #(#args_pat: #args_ty),* ) -> savvy::Result<savvy::SEXP> {
-                    let self__ = savvy::get_external_pointer_addr(self__) as *mut #ty;
-                    #(#stmts_additional)*
-
-                    (*self__).#fn_name_orig(#(#args_pat),*)
-                }
-            ),
-            // A method without result; return a dummy value
-            (SavvyFnType::Method(ty), false) => parse_quote!(
-                #(#attrs)*
-                #[allow(non_snake_case)]
-                unsafe fn #fn_name_inner(self__: savvy::SEXP, #(#args_pat: #args_ty),* ) -> savvy::Result<savvy::SEXP> {
-                    let self__ = savvy::get_external_pointer_addr(self__) as *mut #ty;
-                    #(#stmts_additional)*
-
-                    (*self__).#fn_name_orig(#(#args_pat),*);
-
-                    // Dummy return value
-                    Ok(savvy::NullSxp.into())
-                }
-            ),
-            // An associated function with a result
-            (SavvyFnType::AssociatedFunction(ty), true) => parse_quote!(
-                #(#attrs)*
-                #[allow(non_snake_case)]
-                unsafe fn #fn_name_inner(#(#args_pat: #args_ty),* ) -> savvy::Result<savvy::SEXP> {
-                    #(#stmts_additional)*
-
-                    #ty::#fn_name_orig(#(#args_pat),*)
-                }
-            ),
-            // An associated function without result; return a dummy value
-            (SavvyFnType::AssociatedFunction(ty), false) => parse_quote!(
-                #(#attrs)*
-                #[allow(non_snake_case)]
-                unsafe fn #fn_name_inner(#(#args_pat: #args_ty),* ) -> savvy::Result<savvy::SEXP> {
-                    #(#stmts_additional)*
-
-                    #ty::#fn_name_orig(#(#args_pat),*);
-
-                    // Dummy return value
-                    Ok(savvy::NullSxp.into())
-                }
-            ),
-            (SavvyFnType::Constructor(ty), _) => parse_quote!(
-                #(#attrs)*
-                #[allow(non_snake_case)]
-                unsafe fn #fn_name_inner(#(#args_pat: #args_ty),* ) -> savvy::Result<savvy::SEXP> {
-                    #(#stmts_additional)*
-                    let x = #ty::#fn_name_orig(#(#args_pat),*);
-                    Ok(x.into_external_pointer())
-                }
-            ),
-        };
-        out
-    }
-
-    pub fn generate_outer_fn(&self) -> syn::ItemFn {
-        let fn_name_inner = self.fn_name_inner();
-        let fn_name_outer = self.fn_name_outer();
-
-        let args_pat: Vec<syn::Ident> = self.args.iter().map(|arg| arg.pat.clone()).collect();
-        let args_ty: Vec<syn::Type> = self
-            .args
-            .iter()
-            .map(|arg| arg.ty.to_rust_type_inner())
-            .collect();
-
-        let out: syn::ItemFn = match &self.fn_type {
-            // if the function is a method, add `self__` to the first argument
-            SavvyFnType::Method(_) => parse_quote!(
-                #[allow(clippy::missing_safety_doc)]
-                #[no_mangle]
-                pub unsafe extern "C" fn #fn_name_outer(self__: savvy::SEXP, #(#args_pat: #args_ty),* ) -> savvy::SEXP {
-                    savvy::handle_result(#fn_name_inner(self__, #(#args_pat),*))
-                }
-            ),
-            _ => parse_quote!(
-                #[allow(clippy::missing_safety_doc)]
-                #[no_mangle]
-                pub unsafe extern "C" fn #fn_name_outer( #(#args_pat: #args_ty),* ) -> savvy::SEXP {
-                    savvy::handle_result(#fn_name_inner(#(#args_pat),*))
-                }
-            ),
-        };
-        out
-    }
-
-    fn get_c_args(&self) -> Vec<(String, String)> {
-        let mut out: Vec<_> = self
-            .args
-            .iter()
-            .map(|arg| {
-                let pat = arg.pat.to_string();
-                let ty = arg.ty.to_c_type();
-                (pat, ty)
-            })
-            .collect();
-
-        // if it's a method, add `self__` arg
-        if let SavvyFnType::Method(_) = &self.fn_type {
-            out.insert(0, ("self__".to_string(), "SEXP".to_string()))
-        }
-
-        out
-    }
-
-    fn get_r_args(&self) -> Vec<String> {
-        let mut out: Vec<_> = self.args.iter().map(|arg| arg.pat.to_string()).collect();
-
-        // if it's a method, add `self__` arg
-        if let SavvyFnType::Method(_) = &self.fn_type {
-            out.insert(0, "self".to_string())
-        }
-
-        out
-    }
-
-    /// Generate C function signature
-    fn to_c_function_for_header(&self) -> String {
-        let fn_name = self.fn_name_outer();
-        let args = self
-            .get_c_args()
-            .iter()
-            .map(|(pat, ty)| format!("{ty} {pat}"))
-            .collect::<Vec<String>>()
-            .join(", ");
-
-        format!("SEXP {fn_name}({args});")
-    }
-
-    /// Generate C function implementation
-    fn to_c_function_impl(&self) -> String {
-        let fn_name = self.fn_name_outer();
-        let args = self.get_c_args();
-
-        let args_sig = args
-            .iter()
-            .map(|(pat, ty)| format!("{ty} {pat}"))
-            .collect::<Vec<String>>()
-            .join(", ");
-
-        let args_call = args
-            .iter()
-            .map(|(pat, _)| pat.as_str())
-            .collect::<Vec<&str>>()
-            .join(", ");
-
-        format!(
-            "
-SEXP {fn_name}_wrapper({args_sig}) {{
-    SEXP res = {fn_name}({args_call});
-    return handle_result(res);
-}}"
-        )
-    }
-
-    /// Generate C function call entry
-    fn to_c_function_call_entry(&self) -> String {
-        let fn_name = self.fn_name_outer();
-        let n_args = self.get_c_args().len();
-        format!(r#"    {{"{fn_name}", (DL_FUNC) &{fn_name}_wrapper, {n_args}}},"#)
-    }
-
-    fn to_r_function(&self) -> String {
-        let fn_name = self.fn_name_r();
-        let fn_name_c = self.fn_name_outer();
-
-        let doc_comments = get_r_doc_comment(self.docs.as_slice());
-
-        let args = self
-            .get_c_args()
-            .iter()
-            .map(|(pat, _)| pat.clone())
-            .collect::<Vec<String>>();
-
-        let mut args_call = args.clone();
-        args_call.insert(0, fn_name_c.to_string());
-
-        let args = args.join(", ");
-        let args_call = args_call.join(", ");
-
-        let body = if self.has_result {
-            format!(".Call({args_call})")
-        } else {
-            // If the result is NULL, wrap it with invisible
-            format!("invisible(.Call({args_call}))")
-        };
-
-        format!(
-            "{doc_comments}
-{fn_name} <- function({args}) {{
-  {body}
-}}
-"
-        )
-    }
-}
-
-fn get_r_doc_comment(docs: &[String]) -> String {
-    docs.iter()
-        .map(|doc| format!("#'{doc}"))
-        .collect::<Vec<String>>()
-        .join("\n")
-}
-
-pub fn generate_c_header_file(parsed_result: &ParsedResult) -> String {
-    let bare_fns = parsed_result
-        .bare_fns
-        .iter()
-        .map(|x| x.to_c_function_for_header())
-        .collect::<Vec<String>>()
-        .join("\n");
-
-    let impls = parsed_result
-        .impls
-        .iter()
-        .map(|x| {
-            let fns = x
-                .fns
-                .iter()
-                .map(|x| x.to_c_function_for_header())
-                .collect::<Vec<String>>()
-                .join("\n");
-
-            format!(
-                "\n// methods and associated functions for {}\n{}",
-                x.ty, fns
-            )
-        })
-        .collect::<Vec<String>>()
-        .join("\n");
-
-    format!("{bare_fns}\n{impls}")
-}
-
-fn generate_c_function_impl(fns: &[SavvyFn]) -> String {
-    fns.iter()
-        .map(|x| x.to_c_function_impl())
-        .collect::<Vec<String>>()
-        .join("\n")
-}
-
-fn generate_c_function_call_entry(fns: &[SavvyFn]) -> String {
-    fns.iter()
-        .map(|x| x.to_c_function_call_entry())
-        .collect::<Vec<String>>()
-        .join("\n")
-}
-
-pub fn generate_c_impl_file(parsed_result: &ParsedResult, pkg_name: &str) -> String {
-    let common_part = r#"
-#include <stdint.h>
-#include <Rinternals.h>
-#include "rust/api.h"
-
-static uintptr_t TAGGED_POINTER_MASK = (uintptr_t)1;
-
-SEXP handle_result(SEXP res_) {
-    uintptr_t res = (uintptr_t)res_;
-
-    // An error is indicated by tag.
-    if ((res & TAGGED_POINTER_MASK) == 1) {
-        // Remove tag
-        SEXP res_aligned = (SEXP)(res & ~TAGGED_POINTER_MASK);
-
-        // Currently, there are two types of error cases:
-        //
-        //   1. Error from Rust code
-        //   2. Error from R's C API, which is caught by R_UnwindProtect()
-        //
-        if (TYPEOF(res_aligned) == CHARSXP) {
-            // In case 1, the result is an error message that can be passed to
-            // Rf_error() directly.
-            Rf_error("%s", CHAR(res_aligned));
-        } else {
-            // In case 2, the result is the token to restart the
-            // cleanup process on R's side.
-            R_ContinueUnwind(res_aligned);
-        }
-    }
-
-    return (SEXP)res;
-}
-"#;
-
-    let c_fns_bare = generate_c_function_impl(parsed_result.bare_fns.as_slice());
-
-    let c_fns_impl = parsed_result
-        .impls
-        .iter()
-        .map(|x| {
-            format!(
-                "\n// methods and associated functions for {}\n{}",
-                x.ty,
-                generate_c_function_impl(x.fns.as_slice())
-            )
-        })
-        .collect::<Vec<String>>()
-        .join("\n");
-
-    let call_entries_bare = generate_c_function_call_entry(parsed_result.bare_fns.as_slice());
-
-    let call_entries_impl = parsed_result
-        .impls
-        .iter()
-        .map(|x| {
-            format!(
-                "\n// methods and associated functions for {}\n{}",
-                x.ty,
-                generate_c_function_call_entry(x.fns.as_slice())
-            )
-        })
-        .collect::<Vec<String>>()
-        .join("\n");
-
-    format!(
-        "{common_part}
-{c_fns_bare}
-{c_fns_impl}
-
-static const R_CallMethodDef CallEntries[] = {{
-{call_entries_bare}
-{call_entries_impl}
-    {{NULL, NULL, 0}}
-}};
-
-void R_init_{pkg_name}(DllInfo *dll) {{
-  R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
-  R_useDynamicSymbols(dll, FALSE);
-}}
-"
-    )
-}
-
-fn generate_r_impl_for_impl(savvy_impl: &SavvyImpl) -> String {
-    let mut ctors: Vec<&SavvyFn> = Vec::new();
-    let mut others: Vec<&SavvyFn> = Vec::new();
-    let class_r = savvy_impl.ty.clone();
-
-    for savvy_fn in &savvy_impl.fns {
-        match savvy_fn.fn_type {
-            SavvyFnType::Constructor(_) => ctors.push(savvy_fn),
-            _ => others.push(savvy_fn),
-        }
-    }
-
-    // TODO: error if no ctor
-
-    let closures = others
-        .iter()
-        .map(|x| {
-            let fn_name = x.fn_name_r();
-            let fn_name_c = x.fn_name_outer();
-
-            let mut args = x.get_r_args();
-            // Remove self from arguments for R
-            let args_r = args
-                .clone()
-                .into_iter()
-                .filter(|e| *e != "self")
-                .collect::<Vec<String>>()
-                .join(", ");
-
-            args.insert(0, fn_name_c.to_string());
-            let args_call = args.join(", ");
-
-            let body = if x.has_result {
-                format!(".Call({args_call})")
-            } else {
-                // If the result is NULL, wrap it with invisible
-                format!("invisible(.Call({args_call}))")
-            };
-
-            format!(
-                "{fn_name} <- function(self) {{
-  function({args_r}) {{
-    {body}
-  }}
-}}
-"
-            )
-        })
-        .collect::<Vec<String>>()
-        .join("\n");
-
-    let builders = ctors
-        .iter()
-        .map(|x| {
-            let fn_name = x.fn_name_r();
-            let fn_name_c = x.fn_name_outer();
-
-            let mut args = x.get_r_args();
-
-            let args_r = args.join(", ");
-
-            args.insert(0, fn_name_c.to_string());
-            let args_call = args.join(", ");
-
-            let methods = others
-                .iter()
-                .map(|o| format!("  e${} <- {}(self)", o.fn_name, o.fn_name_r()))
-                .collect::<Vec<String>>()
-                .join("\n");
-
-            format!(
-                r#"{fn_name} <- function({args_r}) {{
-  e <- new.env(parent = emptyenv())
-  self <- .Call({args_call})
-
-{methods}
-
-  class(e) <- "{class_r}"
-  e
-}}
-"#
-            )
-        })
-        .collect::<Vec<String>>()
-        .join("\n");
-
-    let doc_comments = get_r_doc_comment(savvy_impl.docs.as_slice());
-
-    format!(
-        "{doc_comments}
-{builders}
-
-{closures}
-"
-    )
-}
-
-pub fn generate_r_impl_file(parsed_result: &ParsedResult, pkg_name: &str) -> String {
-    let r_fns = parsed_result
-        .bare_fns
-        .iter()
-        .map(|x| x.to_r_function())
-        .collect::<Vec<String>>()
-        .join("\n");
-
-    let r_impls = parsed_result
-        .impls
-        .iter()
-        .map(generate_r_impl_for_impl)
-        .collect::<Vec<String>>()
-        .join("\n");
-
-    format!(
-        r#"#' @useDynLib {pkg_name}, .registration = TRUE
-#' @keywords internal
-NULL
-
-{r_fns}
-{r_impls}
-"#
-    )
 }
